feat: add LearningPathStageListScreen
